move modern TabBar
