+++ src/screens/WalletConnect/Home/WalletConnectHome.js
@@ -30,7 +30,7 @@ import { chunk } from 'lodash';
 // Components
 import { Container, Center } from 'components/layout/Layout';
 import HeaderBlock from 'components/HeaderBlock';
-import TabBar from 'components/modern/TabBar';
+import TabBar from 'components/layout/TabBar';
 import Text from 'components/core/Text';
 import FloatingButtons from 'components/FloatingButtons';
 import Spinner from 'components/Spinner';
